Added Routing logic & fixed bugs in About-us section
