+++ src/App.jsx
@@ -1,30 +1,18 @@
-import AuthPage from "./AuthPage";
-import Aboutus from "./components/Aboutus";
-import ContactUs from "./components/ContactUs";
-import FAQ from "./components/FAQ";
-import Footer from "./components/Footer";
-import Hero from "./components/Hero";
 import NavBar from "./components/NavBar";
-import Programs from "./components/Programs";
-import Testimonial from "./components/Testimonial";
-import Timeline from "./components/Timeline";
+import AuthPage from "./pages/AuthPage";
+import LandingPage from "./pages/LandingPage";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
 
 function App() {
   return (
-    <>
-      <NavBar />
-      <main>
-        {/* <Hero />
-        <Aboutus />
-        <Timeline />
-        <Programs />
-        <Testimonial />
-        <FAQ />
-        <ContactUs /> */}
-        <AuthPage />
-      </main>
-      <Footer />
-    </>
+    <BrowserRouter>
+      <Routes>
+        <Route path="/">
+          <Route index element={<LandingPage />} />
+          <Route path="/auth" element={<AuthPage />} />
+        </Route>
+      </Routes>
+    </BrowserRouter>
   );
 }
